requirements.txt
requests
gradio
openai
Pillow
pytest
